requirements.txt
streamlit
pandas
pypdf
google-generativeai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client